Bisection in Python
The bisection method uses the intermediate value theorem iteratively to find roots. Let f(x) be a continuous function, and a and b be real scalar values such that a<b. Assume, without loss of generality, that f(a)>0 and f(b)<0. Then by the intermediate value theorem, there must be a root on the open interval (a,b).